feat(ACI): Create shared test base class for workflow engine serializers
